references/ADVANCED-transcription/text-to-speech/Trelis-orpheus/fine-tune/requirements.txt
sounddevice
soundfile
numpy
datasets
huggingface_hub
transformers
torch
wandb
trl
accelerate
pyyaml
colorama 